Improve intra-package imports
1) Allow import of DenseROISchema from ophys_etl.schemas

   Previously this required:
   `from ophys_etl.schemas.dense_roi import DenseROISchema`

   Now it is:
   `from ophys_etl.schemas import DenseROISchema`

2) Allow import of DenseROI TypedDict from ophys_etl.types

   Previously this required:
   `from ophys_etl.transforms.roi_transforms import DenseROI`

   Now it is:
   `from ophys_etl.types import DenseROI`
