+++ src/ophys_etl/schemas/__init__.py
@@ -1,0 +1,1 @@
+from ophys_etl.schemas._dense_roi_schema import DenseROISchema  # noqa: F401
