+++ src/ophys_etl/schemas/_dense_roi_schema.py
