+++ src/ophys_etl/transforms/roi_transforms.py
@@ -1,29 +1,8 @@
-import sys
 from typing import List, Optional, Tuple
 import numpy as np
 from scipy.sparse import coo_matrix
 from ophys_etl.extractors.motion_correction import MotionBorder
-
-if sys.version_info >= (3, 8):
-    from typing import TypedDict
-else:
-    from typing_extensions import TypedDict
-
-
-class DenseROI(TypedDict):
-    id: int
-    x: int
-    y: int
-    width: int
-    height: int
-    valid_roi: bool
-    mask_matrix: List[List[bool]]
-    max_correction_up: int
-    max_correction_down: int
-    max_correction_left: int
-    max_correction_right: int
-    mask_image_plane: int
-    exclusion_labels: List[str]
+from ophys_etl.types import DenseROI
 
 
 def suite2p_rois_to_coo(suite2p_stats: np.ndarray,
