+++ src/ophys_etl/types/__init__.py
@@ -1,0 +1,1 @@
+from ophys_etl.types._dense_roi import DenseROI  # noqa: F401
